Fix ScrollInview component to always trigger.
Since it collects the changes to apply, and then applies in bulk during an RAF callback, it was triggering a race condition.

We were only triggering the inview change to happen if the new value is different than the current. However, since we queue up the changes, it wasn't triggering the boolean flip to happen in cases where it flips back and forth really fast.

For example, if you go from false to true and then true to false before the RAF callback happened, the second change wouldn't queue since it is still false at the time the second change tries to happen.